Fix warnings about precision loss from long to int
Kind of:
pg_connection.c:706:17: warning: implicit conversion loses integer precision: 'long' to 'int' [-Wshorten-64-to-32]
        return INT2NUM(atol(port));

These warnings were emitted on Macos.

+++ ext/pg_connection.c
@@ -703,7 +703,7 @@ static VALUE
 pgconn_port(VALUE self)
 {
 	char* port = PQport(pg_get_pgconn(self));
-	return INT2NUM(atol(port));
+	return INT2NUM(atoi(port));
 }
 
 /*
